+++ Makefile
@@ -1,6 +1,5 @@
 SRC_DIR=src
 FRONTEND_DIR=$(SRC_DIR)/frontend/infrastructure/discord
-DB_NAME=papers.db
 MIGRATIONS_DIR=migrations
 
 MIGRATION_FILES=$(shell find $(MIGRATIONS_DIR) -name '*.sql' | sort)
@@ -8,10 +7,10 @@ MIGRATION_FILES=$(shell find $(MIGRATIONS_DIR) -name '*.sql' | sort)
 .PHONY: run migrate lint
 
 run:
-	cd $(FRONTEND_DIR) && deno run --allow-net --allow-read --allow-env --allow-write bot.ts
+    cd $(FRONTEND_DIR) && deno run --allow-net --allow-read --allow-env --allow-write bot.ts
 
 migrate:
-	@$(foreach file,$(MIGRATION_FILES),sqlite3 $(DB_NAME) < $(file);)
+    @$(foreach file,$(MIGRATION_FILES),mysql -u$(MYSQL_USERNAME) -p$(MYSQL_PASSWORD) $(DATABASE_NAME) < $(file);)
 
 lint:
-	deno lint
+    deno lint
